Add files via upload
Here I stored my Data Structure workings that are included Stacks, Queues, Linked Lists by using both arrays and linked lists.
Index of codes and how it works is included in files.

In turtle.cpp :
The Logo language, which is particularly popular among personal computer users, made the concept
of turtle graphics famous. Imagine a mechanical turtle that walks around the room under the control
of a C++ program.
1 Pen up
2 Pen down
3 Turn right
4 Turn left
5 10 Move forward 10 spaces (or a number other than 10)
6  Print the 20-by-20 array
9  End of data
--------------------------------------------------------------------------------------------------------------------------------------------------------------
In homework2.c :
we have 4 different classes of Data Structures. The number of students in these classes may be different.
 Leave student list as sorted when performing the insert operation. Sort will be performed in decreasing order based on the midterm scores. And the program calculates the averages.
------------------------------------------------------------------------------------------------------------------------------------------------------------------------
In homework3:
We'll see how to combine two different data structure Linked List  in one Linked List.

--------------------------------------------------------------------------------------------------------------------------------------------------------------------------
In homework4:
This is an Employee-Customer Banking computer simulation system.You go to bank and start to wait in queue. Whenever an employee is available this is your turn.We'll see how this system works.The output will consist of customer number,
employee number, time the transaction started, time the transaction finished, customer waiting time in queue.
Additionally, each employee's total serving time will be printed. Write and use all the necessary methods of "Stack"
and "Queue" structures.
------------------------------------------------------------------------------------------------------------------------------------------------------------------
In homework5:
Student IDs and grades are given in input.We'll see how binary search tree works.I also provided an simple image of the created tree in this program as an output.
-------------------------------------------------------------------------------------------------------------------------------------------------------------------
In homework6:
In this homework we've implemented an AVL tree.Output will be an inorder walk on created AVL tree.
-------------------------------------------------------------------------------------------------------------------------------------------------------------------